+++ modules/edit-modes/src/lib/draw-circle-by-diameter-mode.js
@@ -3,9 +3,10 @@
 import circle from '@turf/circle';
 import distance from '@turf/distance';
 import type { Position, Polygon, FeatureOf } from '../geojson-types.js';
+import { getIntermediatePosition } from './geojson-edit-mode.js';
 import { TwoClickPolygonMode } from './two-click-polygon-mode.js';
 
-export class DrawCircleFromCenterMode extends TwoClickPolygonMode {
+export class DrawCircleByDiameterMode extends TwoClickPolygonMode {
   getTwoClickPolygon(coord1: Position, coord2: Position, modeConfig: any): FeatureOf<Polygon> {
     // Default turf value for circle is 64
     const { steps = 64 } = modeConfig || {};
@@ -16,8 +17,9 @@ export class DrawCircleFromCenterMode extends TwoClickPolygonMode {
       options.steps = 4;
     }
 
-    const radius = Math.max(distance(coord1, coord2), 0.001);
+    const centerCoordinates = getIntermediatePosition(coord1, coord2);
+    const radius = Math.max(distance(coord1, centerCoordinates), 0.001);
 
-    return circle(coord1, radius, options);
+    return circle(centerCoordinates, radius, options);
   }
 }