+++ modules/edit-modes/src/lib/two-click-polygon-mode.js
@@ -1,15 +1,51 @@
 // @flow
 
-import type { ClickEvent, ModeProps } from '../types.js';
-import type { FeatureCollection } from '../geojson-types.js';
-import { BaseGeoJsonEditMode, type GeoJsonEditAction } from './geojson-edit-mode.js';
+import type {
+  ClickEvent,
+  StartDraggingEvent,
+  StopDraggingEvent,
+  PointerMoveEvent,
+  ModeProps,
+  GuideFeatureCollection
+} from '../types.js';
+import type { Polygon, FeatureCollection, FeatureOf, Position } from '../geojson-types.js';
+import { BaseGeoJsonEditMode } from './geojson-edit-mode.js';
 
 export class TwoClickPolygonMode extends BaseGeoJsonEditMode {
-  handleClickAdapter(event: ClickEvent, props: ModeProps<FeatureCollection>): ?GeoJsonEditAction {
-    super.handleClickAdapter(event, props);
+  handleClick(event: ClickEvent, props: ModeProps<FeatureCollection>) {
+    if (props.modeConfig && props.modeConfig.dragToDraw) {
+      // handled in drag handlers
+      return;
+    }
 
-    const tentativeFeature = this.getTentativeFeature();
+    this.addClickSequence(event);
+
+    this.checkAndFinishPolygon(props);
+  }
+
+  handleStartDragging(event: StartDraggingEvent, props: ModeProps<FeatureCollection>): void {
+    if (!props.modeConfig || !props.modeConfig.dragToDraw) {
+      // handled in click handlers
+      return;
+    }
+
+    this.addClickSequence(event);
+    event.cancelPan();
+  }
+
+  handleStopDragging(event: StopDraggingEvent, props: ModeProps<FeatureCollection>): void {
+    if (!props.modeConfig || !props.modeConfig.dragToDraw) {
+      // handled in click handlers
+      return;
+    }
+    this.addClickSequence(event);
+
+    this.checkAndFinishPolygon(props);
+  }
+
+  checkAndFinishPolygon(props: ModeProps<FeatureCollection>) {
     const clickSequence = this.getClickSequence();
+    const tentativeFeature = this.getTentativeGuide(props);
 
     if (
       clickSequence.length > 1 &&
@@ -17,15 +53,51 @@ export class TwoClickPolygonMode extends BaseGeoJsonEditMode {
       tentativeFeature.geometry.type === 'Polygon'
     ) {
       const editAction = this.getAddFeatureOrBooleanPolygonAction(tentativeFeature.geometry, props);
+
       this.resetClickSequence();
-      this._setTentativeFeature(null);
-      return editAction;
+
+      if (editAction) {
+        props.onEdit(editAction);
+      }
+    }
+  }
+
+  getGuides(props: ModeProps<FeatureCollection>): GuideFeatureCollection {
+    const { lastPointerMoveEvent, modeConfig } = props;
+    const clickSequence = this.getClickSequence();
+
+    const guides: GuideFeatureCollection = {
+      type: 'FeatureCollection',
+      features: []
+    };
+
+    if (clickSequence.length === 0) {
+      // nothing to do yet
+      return guides;
     }
 
+    const corner1 = clickSequence[0];
+    const corner2 = lastPointerMoveEvent.mapCoords;
+
+    const polygon = this.getTwoClickPolygon(corner1, corner2, modeConfig);
+    if (polygon) {
+      guides.features.push({
+        type: 'Feature',
+        properties: {
+          guideType: 'tentative'
+        },
+        geometry: polygon.geometry
+      });
+    }
+
+    return guides;
+  }
+
+  getTwoClickPolygon(coord1: Position, coord2: Position, modeConfig: any): ?FeatureOf<Polygon> {
     return null;
   }
 
-  getCursorAdapter() {
-    return 'cell';
+  handlePointerMove(event: PointerMoveEvent, props: ModeProps<FeatureCollection>) {
+    props.onUpdateCursor('cell');
   }
 }