+++ website/gatsby-config.js
@@ -1,14 +1,9 @@
-const { getGatsbyConfig } = require('ocular-gatsby/api');
-const config = require('./ocular-config');
+const ocularConfig = require('./ocular-config');
 
-const gatsbyConfig = getGatsbyConfig(config);
-gatsbyConfig.plugins.push('gatsby-plugin-flow');
-
-const remark = gatsbyConfig.plugins.find(p => p.resolve === 'gatsby-transformer-remark');
 // If you are using gatsby-remark-responsive-iframe, it must appear after gatsby-remark-embedded-codesandbox
 // plugin in your configuration or the iframe will not be transformed.
 
-remark.options.plugins.unshift({
+const embeddedCodesandbox = {
   resolve: 'gatsby-remark-embedded-codesandbox',
   options: {
     directory: `${__dirname}/../examples/codesandbox`,
@@ -23,6 +18,22 @@ remark.options.plugins.unshift({
     getIframe: url =>
       `<iframe src="${url}" style="width: 70vw; height: 70vh;" class="embedded-codesandbox" sandbox="allow-modals allow-forms allow-popups allow-scripts allow-same-origin"></iframe>`
   }
-});
+};
 
-module.exports = gatsbyConfig;
+module.exports = {
+  plugins: [
+    {
+      resolve: `gatsby-theme-ocular`,
+      options: ocularConfig
+    },
+    {
+      resolve: `gatsby-plugin-flow`
+    },
+    {
+      resolve: `gatsby-transformer-remark`,
+      options: {
+        plugins: [embeddedCodesandbox]
+      }
+    }
+  ]
+};